+++ ChangeLog.md
@@ -3,6 +3,11 @@
 This changelog is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Added
+- Added localized strings for FR-FR and DE-DE. Translations done from English using CoPilot.
+
+### Removed
+- Removed files previously marked as deprecated
 
 ## [2.1.0] - 2024-11-21